requirements.txt
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.5
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
numpy==1.26.4
notebook==7.2.2
notebook_shim==0.2.4
openpyxl==3.1.2
pandas==2.2.2
XlsxWriter==3.2.0
snowflake-connector-python==3.12.2
environs==10.3.0